+++ app/src/main/java/com/example/mf4z/fulfulde/ColorsActivity.java
@@ -13,135 +13,19 @@ import java.util.ArrayList;
 
 public class ColorsActivity extends AppCompatActivity {
 
-    //Declaring Media Player variable
-    private MediaPlayer mMediaPlayer;
 
-    //Delaring Audio Manager for use in the entire class
-    private AudioManager mAudioManager;
-
-    private MediaPlayer.OnCompletionListener mCompletionListener = new MediaPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion(MediaPlayer mediaPlayer) {
-            // Now that the sound file has finished playing, release the media player resources.
-            releaseMediaPlayer();
-        }
-    };
-
-    private AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener = new AudioManager.OnAudioFocusChangeListener() {
-        @Override
-        public void onAudioFocusChange(int focusChange) {
-            if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT ||
-                    focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK) {
-                // The AUDIOFOCUS_LOSS_TRANSIENT case means that we've lost audio focus for a
-                // short amount of time. The AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK case means that
-                // our app is allowed to continue playing sound but at a lower volume. We'll treat
-                // both cases the same way because our app is playing short sound files.
-
-                // Pause playback and reset player to the start of the file. That way, we can
-                // play the word from the beginning when we resume playback.
-                mMediaPlayer.pause();
-                mMediaPlayer.seekTo(0);
-            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-                // The AUDIOFOCUS_GAIN case means we have regained focus and can resume playback.
-                mMediaPlayer.start();
-            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                // The AUDIOFOCUS_LOSS case means we've lost audio focus and
-                // Stop playback and clean up resources
-                releaseMediaPlayer();
-            }
-        }
-    };
 
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.word_list);
+        setContentView(R.layout.activity_category);
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.container, new ColorsFragment())
+                .commit();
 
-        //Create and Setup
-        mAudioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
-
-        final ArrayList<Word> colors = new ArrayList<>();
-
-
-        colors.add(new Word("Balejum","Black",R.drawable.color_black,R.raw.color_black));
-        colors.add(new Word("Danejum","White",R.drawable.color_white,R.raw.color_white));
-        colors.add(new Word("Bodejum","Red",R.drawable.color_red,R.raw.color_red));
-        colors.add(new Word("Haako Haako","Green",R.drawable.color_green,R.raw.color_green));
-        colors.add(new Word("Ndiyam Goro","Yellow",R.drawable.color_mustard_yellow,R.raw.color_yellow));
-        colors.add(new Word("Purum  Purum","Gray",R.drawable.color_gray,R.raw.color_gray));
-
-
-        WordAdapter adapter = new WordAdapter(this,colors,R.color.colorOrange);
-
-        ListView listView = findViewById(R.id.list);
-
-        listView.setAdapter(adapter);
-
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-
-                // Release the media player if it currently exists because we are about to
-                // play a different sound file
-                releaseMediaPlayer();
-
-                // Get the {@link Word} object at the given position the user clicked on
-                Word word = colors.get(position);
-
-                //Log.v("NumbersActivity", "Current word: " + word);
-
-// Request audio focus for playback
-                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
-                        // Use the music stream.
-                        AudioManager.STREAM_MUSIC,
-                        // Request permanent focus.
-                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
-
-                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-
-                    // We have audio focus now
-
-                    //Creating a media player with a song
-                    mMediaPlayer = MediaPlayer.create(ColorsActivity.this,word.getMwordSound());
-                    //play sound
-                    mMediaPlayer.start();
-
-                    // Setup a listener on the media player, so that we can stop and release the
-                    // media player once the sound has finished playing.
-                    mMediaPlayer.setOnCompletionListener(mCompletionListener);
-
-                }
-
-            }
-        });
 
     }
 
-    @Override
-    protected void onStop() {
-        super.onStop();
-        releaseMediaPlayer();
-    }
 
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer() {
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-
-            // Regardless of whether or not we were granted audio focus, abandon it. This also
-            // unregisters the AudioFocusChangeListener so we don't get anymore callbacks.
-            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
-        }
-    }
 }
